Updated Firebase Attempt
