+++ sai-challenge4/public/index.html
@@ -21,7 +21,10 @@
     -->
     <title>React App</title>
   </head>
-  <script src="https://www.gstatic.com/firebasejs/4.1.0/firebase.js"></script>
+  <body>
+    <div id="root"></div>
+
+     <script src="https://www.gstatic.com/firebasejs/4.1.0/firebase.js"></script>
 <script>
   // Initialize Firebase
   var config = {
@@ -33,21 +36,6 @@
     messagingSenderId: "632497945652"
   };
   firebase.initializeApp(config);
-</script>
-  <body>
-    <noscript>
-      You need to enable JavaScript to run this app.
-    </noscript>
-    <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
+  </script>
   </body>
 </html>
